blog_writer_app/requirements.txt
streamlit>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
typing-extensions>=4.7.0
python-dotenv>=1.0.0
nltk>=3.8.1
